+++ packages/aimd-recorder/src/components/index.ts
@@ -10,8 +10,9 @@ import AimdRecorder from "./AimdRecorder.vue"
 import AimdQuizRecorder from "./AimdQuizRecorder.vue"
 
 const AimdDnaSequenceField = defineAsyncComponent(() => import("./AimdDnaSequenceField.vue"))
+const AimdMarkdownField = defineAsyncComponent(() => import("./AimdMarkdownField.vue"))
 
-export { AimdRecorder, AimdQuizRecorder, AimdDnaSequenceField }
+export { AimdRecorder, AimdQuizRecorder, AimdDnaSequenceField, AimdMarkdownField }
 
 /**
  * @deprecated Use `AimdRecorder` instead.